requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
google-genai>=0.1.0
pydantic>=2.0.0
pytest>=7.0.0
httpx>=0.24.0
